Changed `DatabaseHandler` to be an interface, renamed the implementation to `RealmDatabaseHandler`
